add: mvideo parser, handle lazy load

+++ backend/app/dto/products.py
@@ -3,7 +3,7 @@ from pydantic import BaseModel
 
 class BaseProductModel(BaseModel):
     name: str
-    oldPrice: float | None = None
+    oldPrice: int | None = None
     price: int
     bonus: int | None = None
     url: str
